Rotation Algorithm Refactoring
Changed the algorithm used.
Makes a new vec and inserts in a different order to get a rotated vec
It's a bit more readable and would extend to larger shapes better. Or
rectangles instead of squares.
It does make a new vec and replace the old one instead of swapping in
place. Not that that should matter unless the shapes are 10000x100000. And
it's inevitable if the intention is to make a new shape instead of
changing the existing shape.

Gave the option to return a new shape instead of changing the existing
shape. I think that's what the exercise wants?